Nest sync jobs under namespace for the sake of clarity
