+++ app/jobs/generate_attributive_decision_job.rb
@@ -15,7 +15,7 @@ class GenerateAttributiveDecisionJob < ApplicationJob
   end
 
   def perform(schooling)
-    FetchStudentInformationJob.new.perform(schooling) if schooling.student.missing_address?
+    Sync::StudentJob.new.perform(schooling) if schooling.student.missing_address?
 
     Schooling.transaction do
       generate_document(schooling)
