+++ app/jobs/prepare_payment_request_job.rb
@@ -4,7 +4,7 @@ class PreparePaymentRequestJob < ApplicationJob
   include FregataProof
 
   def perform(payment_request)
-    FetchStudentInformationJob.perform_now(payment_request.schooling)
+    Sync::StudentJob.perform_now(payment_request.schooling)
 
     payment_request.mark_ready!
   end
